sapcpos/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Werkzeug==3.0.3
SQLAlchemy==2.0.31
python-dotenv==1.0.1
bcrypt==4.2.1
requests==2.32.3
gunicorn==22.0.0
itsdangerous==2.2.0
